Living Docks Phase 2: live money economy + multi-cell holding
STEP A -- money-width fix (foundation correctness):
- ItemsLiteEntry.quantity short -> int (a vault COIN stack counts a whole
  district's Royals; short's 32767 ceiling clamped and destroyed specie).
- ACTR format: quantity write/readShort -> write/readInt.
- addCarried/addOnCell now return the amount actually added (overflow-guarded);
  moveCarried/moveCarriedToCell/moveCellToCarried return the DEST-ADDED amount
  (a deposit credits Royals by this, so it must equal what landed).
- BankLedger.credit/transfer reject before an int/long overflow wraps.

STEP B -- bank go-live (Pass 9), closed supply, all minting at bake:
- CivicAccounts seeds per-actor balances by wealth tier, mints one ID card per
  citizen, funds a finite EMPLOYER pool account, and stocks the vault chest so
  totalRoyals() == vault COIN count (asserted at bake).
- New sim-core seams (baked, purity-clean int[]/long[]): BankQueue (deterministic
  ascending-id slot assignment, absolute-tick service), Payroll (wage table +
  employer pool), CivicFixtures (bundles arrest cell/zones/bank/prison/payroll),
  wired via a new ActorsSystem constructor; ActorContext exposes vaultChestCell/
  bankerCell/bankQueue/prisonCells (defaults NONE/EMPTY).
- Live wages: ActorsSystem.tick transfers each wage employer->worker on a fixed
  cadence, ascending id -- a transfer, never a mint (insufficient pool skips).
- BankVerbs.depositCoins/withdrawRoyals/buyFood: the only Coin<->Royal on/off-ramp
  and an ID-authorized Royal purchase that refuses loose Coins + insufficient
  balance (callable + unit-tested; behavior triggers land in Phase 3).

STEP C -- multi-cell holding (Pass 10):
- Actor gains a serialized assignedHoldCell scalar (serialize/load/hashInto).
- PrisonCellRegistry wired from prisonCellsK34(); arrest assigns the least-crowded
  free cell (ascending, respects MAX_OCCUPANTS_PER_CELL=2) so six prisoners hold
  in six distinct cells; HeldPolicy escorts there and release frees the slot.

Verified: gradlew build --rerun-tasks BUILD SUCCESSFUL (ArchUnit + all tests);
ACTR round-trip byte-identical (int quantity + seeded ledger + assigned cell);
two runDocksActors --ticks 20000 runs byte-identical (SHA c3dd5303...);
640 actors, 641 accounts; totalRoyals==vaultCOIN and COIN supply conserved after
20k and 120k (5-day, ~20-payday) runs.

+++ sim-core/src/main/java/com/trojia/sim/actor/ItemsLiteEntry.java
@@ -19,13 +19,17 @@ package com.trojia.sim.actor;
  * @param locationCarriedBy  carrying actor id, or {@link Actor#NONE}
  * @param locationCell       cell it sits on, or {@link Actor#NONE} — mutually
  *                          exclusive with {@code locationCarriedBy}
- * @param quantity           stack count; {@code >= 1}
+ * @param quantity           stack count; {@code >= 1}. An {@code int} (not a
+ *                          {@code short}): a vault COIN stack holds a whole
+ *                          district's Royals ({@link BankLedger#totalRoyals()}),
+ *                          which a {@code short}'s 32767 ceiling would clamp and
+ *                          silently destroy specie (Phase-2 STEP A money-width fix)
  * @param accountId          bank account this item authorizes (ID_CARD), else
  *                          {@link Actor#NONE}
  */
 public record ItemsLiteEntry(
         int itemId, short kindId, int ownerActorId,
-        int locationCarriedBy, int locationCell, short quantity, int accountId) {
+        int locationCarriedBy, int locationCell, int quantity, int accountId) {
 
     public ItemsLiteEntry {
         if (locationCarriedBy != Actor.NONE && locationCell != Actor.NONE) {
@@ -43,7 +47,7 @@ public record ItemsLiteEntry(
 
     /** Convenience for the common non-account item (stamps {@link Actor#NONE}). */
     public ItemsLiteEntry(int itemId, short kindId, int ownerActorId,
-            int locationCarriedBy, int locationCell, short quantity) {
+            int locationCarriedBy, int locationCell, int quantity) {
         this(itemId, kindId, ownerActorId, locationCarriedBy, locationCell, quantity, Actor.NONE);
     }
 }
